an alternative transaction validation approach

+++ base_layer/wallet/migrations/2019-10-30-084148_output_manager_service/up.sql
@@ -8,7 +8,7 @@ CREATE TABLE outputs (
 );
 
 CREATE TABLE pending_transaction_outputs (
-    tx_id INTEGER PRIMARY KEY NOT NULL,
+    tx_id BIGINT PRIMARY KEY NOT NULL,
     short_term INTEGER NOT NULL,
     timestamp DATETIME NOT NULL
 );
