+++ base_layer/wallet/migrations/2019-11-20-090620_transaction_service/up.sql
@@ -1,35 +1,35 @@
 CREATE TABLE outbound_transactions (
-    tx_id INTEGER PRIMARY KEY NOT NULL,
+    tx_id BIGINT PRIMARY KEY NOT NULL,
     destination_public_key BLOB NOT NULL,
-    amount INTEGER NOT NULL,
-    fee INTEGER NOT NULL,
+    amount BIGINT NOT NULL,
+    fee BIGINT NOT NULL,
     sender_protocol TEXT NOT NULL,
     message TEXT NOT NULL,
     timestamp DATETIME NOT NULL
 );
 
 CREATE TABLE inbound_transactions (
-    tx_id INTEGER PRIMARY KEY NOT NULL,
+    tx_id BIGINT PRIMARY KEY NOT NULL,
     source_public_key BLOB NOT NULL,
-    amount INTEGER NOT NULL,
+    amount BIGINT NOT NULL,
     receiver_protocol TEXT NOT NULL,
     message TEXT NOT NULL,
     timestamp DATETIME NOT NULL
 );
 
 CREATE TABLE coinbase_transactions (
-    tx_id INTEGER PRIMARY KEY NOT NULL,
-    amount INTEGER NOT NULL,
+    tx_id BIGINT PRIMARY KEY NOT NULL,
+    amount BIGINT NOT NULL,
     commitment BLOB NOT NULL,
     timestamp DATETIME NOT NULL
 );
 
 CREATE TABLE completed_transactions (
-    tx_id INTEGER PRIMARY KEY NOT NULL,
+    tx_id BIGINT PRIMARY KEY NOT NULL,
     source_public_key BLOB NOT NULL,
     destination_public_key BLOB NOT NULL,
-    amount INTEGER NOT NULL,
-    fee INTEGER NOT NULL,
+    amount BIGINT NOT NULL,
+    fee BIGINT NOT NULL,
     transaction_protocol TEXT NOT NULL,
     status INTEGER NOT NULL,
     message TEXT NOT NULL,
